add kako login

+++ pages/api/users/confirm.ts
@@ -7,27 +7,36 @@ async function handler(
   req: NextApiRequest,
   res: NextApiResponse<ResponseType>
 ) {
-  const { token } = req.body;
-  const foundToken = await client.token.findUnique({
-    where: {
-      token,
-    },
-    include: {
-      user: true,
-    },
-  });
+  const { token, kakao, kakaoId } = req.body;
+  if (token) {
+    const foundToken = await client.token.findUnique({
+      where: {
+        token,
+      },
+      include: {
+        user: true,
+      },
+    });
+    if (!foundToken) return res.status(404).end();
+    req.session.user = {
+      id: foundToken?.userId,
+    };
+    await req.session.save();
+    await client.token.deleteMany({
+      where: {
+        userId: foundToken?.userId,
+      },
+    });
+    res.json({ ok: true });
+  }
 
-  if (!foundToken) return res.status(404).end();
-  req.session.user = {
-    id: foundToken?.userId,
-  };
-  await req.session.save();
-  await client.token.deleteMany({
-    where: {
-      userId: foundToken.userId,
-    },
-  });
-  res.json({ ok: true });
+  if (kakao) {
+    req.session.user = {
+      id: kakaoId,
+    };
+    await req.session.save();
+    res.json({ ok: true });
+  }
 }
 
 export default withApiSession(withHandler({ methods: ["POST"], handler }));
